+++ packages/orq-rc/src/models/components/evaluatorresponsehttp.ts
@@ -46,8 +46,8 @@ export const EvaluatorResponseHttp$inboundSchema: z.ZodType<
 > = z.object({
   _id: z.string(),
   description: z.string(),
-  created: z.string().default("2026-08-20T04:17:08.353Z"),
-  updated: z.string().default("2026-08-20T04:17:08.353Z"),
+  created: z.string().default("2026-08-20T11:48:22.539Z"),
+  updated: z.string().default("2026-08-20T11:48:22.539Z"),
   updated_by_id: z.nullable(z.string()).optional(),
   project_id: z.string().optional(),
   guardrail_config: z.any().optional(),