+++ packages/orq-rc/src/models/operations/createdataset.ts
@@ -138,7 +138,7 @@ export const CreateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-08-20T04:17:06.109Z",
+    "2026-08-20T11:48:20.273Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {